Add support for RSP receive (`recv()`) with timeout
This commit add a new (optional) `timeout` parameter to `RSP.recv()`.
When used, `recv()` waits for packet and raises `TimeoutError` when
no data are available till timeout expires.

This will be used later to allow target polling when the stub is idle
(either target is running or there's no interaction from GDB).

Technically, this is done by making input stream of `IOPipe` non-blocking
and then using `select()`-like interface to check for events. Also,
use buffering on socket stream.

All this logic is hidden in `IOPipe` - this works, but if we ever have
targets that can send events (and therefore do not require polling), we'd
have to move this up to `Stub` class.