Add support for (poor mans) soft-breakpoints for Microwatt
This commit adds a very hacky support for software breakpoints. It
*DOES NOT* use standard POWER `trap` instruction but rather uses
`b 0` (`0x48000000`) - that is "jump to itself". This effectively causes
the core to get trapped in an endless loop.

Stub then periodically check the status of the core and if it finds out
the code is spinning in one of the planted "breakpoint" instruction,
it stops the core and report `SIGTRAP` back to GDB.