Do not try to import `urjtag` unless absolutely necessary
This commit changes the code so that it does not require `urjtag` unless
one actually wants to connect to an actual board. This simplifies the
setup for testing and other usecases where no real board is used (such
as [1]).

With this test, running tests with `pytest` works without need for
URJTAG.

[1]: https://github.com/janvrany/pygdbstub/pull/5#issuecomment-2537873922