Add Github actions config to run tests and style checks
This commit adds config for GithubActions that runs tests and style
checks on push to master and on pull requests.

Style checks run on CI are run via pre-commit to make sure the same
set of checks that is run on CI can be run by user (via pre-commit,
see README.md).

Co-authored-by: xuxingliang <xuxingliang@xiaomi.com>